[Huobi] Disable private API integration tests
Split tests between private and public API;
Remove broken dependency.
